+++ internal/cloud/aws_test.go
@@ -2,6 +2,7 @@ package cloud_test
 
 import (
 	"bytes"
+	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -244,8 +245,11 @@ func TestAWSCloud_Send(t *testing.T) {
 							UploadId: aws.String("UPLOAD123"),
 						}, nil
 					},
-					mockUploadMultipartPart: func(*glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
-						return &glacier.UploadMultipartPartOutput{}, nil
+					mockUploadMultipartPart: func(u *glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
+						hash := glacier.ComputeHashes(u.Body)
+						return &glacier.UploadMultipartPartOutput{
+							Checksum: aws.String(hex.EncodeToString(hash.TreeHash)),
+						}, nil
 					},
 					mockCompleteMultipartUpload: func(*glacier.CompleteMultipartUploadInput) (*glacier.ArchiveCreationOutput, error) {
 						return &glacier.ArchiveCreationOutput{
@@ -323,13 +327,16 @@ func TestAWSCloud_Send(t *testing.T) {
 					},
 					mockUploadMultipartPart: func() func(*glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
 						var i int
-						return func(*glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
+						return func(u *glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
 							i++
 							if i >= 5 {
 								return nil, errors.New("part rejected")
-							} else {
-								return &glacier.UploadMultipartPartOutput{}, nil
 							}
+
+							hash := glacier.ComputeHashes(u.Body)
+							return &glacier.UploadMultipartPartOutput{
+								Checksum: aws.String(hex.EncodeToString(hash.TreeHash)),
+							}, nil
 						}
 					}(),
 				},
@@ -340,6 +347,43 @@ func TestAWSCloud_Send(t *testing.T) {
 				},
 			},
 			expectedError: errors.New("error sending an archive part (400). details: part rejected"),
+		},
+		{
+			description: "it should detect when backup part checksum don't match",
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test-")
+				if err != nil {
+					t.Fatalf("error creating file. details: %s", err)
+				}
+				defer f.Close()
+
+				f.WriteString(strings.Repeat("Important information for the test backup\n", 1000))
+				return f.Name()
+			}(),
+			multipartUploadLimit: 1024,
+			partSize:             100,
+			awsCloud: cloud.AWSCloud{
+				AccountID: "account",
+				VaultName: "vault",
+				Glacier: glacierAPIMock{
+					mockInitiateMultipartUpload: func(*glacier.InitiateMultipartUploadInput) (*glacier.InitiateMultipartUploadOutput, error) {
+						return &glacier.InitiateMultipartUploadOutput{
+							UploadId: aws.String("UPLOAD123"),
+						}, nil
+					},
+					mockUploadMultipartPart: func(*glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
+						return &glacier.UploadMultipartPartOutput{
+							Checksum: aws.String("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"),
+						}, nil
+					},
+				},
+				Clock: fakeClock{
+					mockNow: func() time.Time {
+						return time.Date(2016, 12, 27, 8, 14, 53, 0, time.UTC)
+					},
+				},
+			},
+			expectedError: errors.New("error comparing checksums on archive part (0)"),
 		},
 		{
 			description: "it should detect an error while completing a big backup upload",
@@ -364,8 +408,11 @@ func TestAWSCloud_Send(t *testing.T) {
 							UploadId: aws.String("UPLOAD123"),
 						}, nil
 					},
-					mockUploadMultipartPart: func(*glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
-						return &glacier.UploadMultipartPartOutput{}, nil
+					mockUploadMultipartPart: func(u *glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
+						hash := glacier.ComputeHashes(u.Body)
+						return &glacier.UploadMultipartPartOutput{
+							Checksum: aws.String(hex.EncodeToString(hash.TreeHash)),
+						}, nil
 					},
 					mockCompleteMultipartUpload: func(*glacier.CompleteMultipartUploadInput) (*glacier.ArchiveCreationOutput, error) {
 						return nil, errors.New("backup too big")
@@ -402,8 +449,11 @@ func TestAWSCloud_Send(t *testing.T) {
 							UploadId: aws.String("UPLOAD123"),
 						}, nil
 					},
-					mockUploadMultipartPart: func(*glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
-						return &glacier.UploadMultipartPartOutput{}, nil
+					mockUploadMultipartPart: func(u *glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
+						hash := glacier.ComputeHashes(u.Body)
+						return &glacier.UploadMultipartPartOutput{
+							Checksum: aws.String(hex.EncodeToString(hash.TreeHash)),
+						}, nil
 					},
 					mockCompleteMultipartUpload: func(*glacier.CompleteMultipartUploadInput) (*glacier.ArchiveCreationOutput, error) {
 						return &glacier.ArchiveCreationOutput{